These changes adress Issue #29 and Issue #30. For issue #29, the maximum C:N for the plant part was used to calculation the lignin:N ratio when plant C / plant N > maximum C:N.
For issue #30, we commented out the density dependent modifications
to the microbial turnover rates (tauR and tauK) because they were
enormous for some low productivity sites.

Also modified the way a point is selected for point-level I/O
in the POINT_casa.csv and POINT_mimics.csv files. In the
fcasacnp_clm_testbed.lst, use the npt index in the restart files.
Previously we used the met.nc index but that was difficult to determine.
These POINT .csv files are used for debugging a cell in the gridded run,
or for output in point-level testbed runs.